Enforce unique items per sport
If a user attempts to create a new item for a given sport, and an item
by the same name already exists for that sport, the item will not be
added to the database. The user remains on the 'new_item' page, and a
flash message informs the user of the duplicate item.