Enable item deletion
Items may now be deleted from the database. The new view function,
'delete_item', displays the new 'DeleteItemForm', which consists of a
single submit field labeled 'Delete'. This form appears in the new
'delete_item' template, which asks the user for confirmation and
displays an anchor (pointing back to the 'item' template) and the submit
button. Upon form submission, the item is deleted from the database, a
flash message informs the user that the item has been deleted, and the
user is redirected to 'home'.