Add GitHub OAuth2 provision
Users may now log in with GitHub. To accomplish this, a new package
called 'dance' has been created. Within the package, the 'utils.py' file
creates a blueprint using the 'flask_dance' 'make_github_blueprint'
method, which takes a client ID and client secret, both of which were
obtained from GitHub developer settings and set to environment
variables. In 'item_catalog/__init__.py', an instance of 'flask_login'
'LoginManager' was created, called 'login_manager'. The 'utils.py' file
imports this and sets the 'user_loader' callback per 'flask-login' and
'flask-dance' documentation. The 'utils.py' file also sets up the
SQLAlchemy backend for multi-user functionality. The 'utils.py' file
sets up the 'github_logged_in' function, which checks for a token and
then retrieves user information to log in the user. After the function
receives an 'OK' response, the function checks the database for the
user's information. If the user is found in the database, the function
logs in the user. If the user is not found in the database, the function
creates a new instance of the user in the database and logs in the user.
Finally, 'utils.py' provides error messages if GitHub sends any errors.
To accomodate these changes, a 'logout' view function has been added to
'main/views.py', and the 'base' template will display a 'Log In with
GitHub' anchor that points to 'github.login' if the user is not logged
in. Otherwise, 'base.html' displays the 'New Item' anchor and 'Log Out'
anchor as usual, except that the 'Log Out' anchor now points to
'main.logout'.