Add custom error pages
Error pages include 404, 403, and 500 errors. To accomplish this, the 
'errors' package was created, containing 'handlers.py'. Within 
'handlers.py', the 'errors' blueprint is created. This blueprint is 
registered in 'item_catalog/__init__.py'. Each error code has a 
corresponding template, all of which are stored in 'templates/errors', 
and each of which features an anchor pointing to 'main.home'.